commit change api response Province, add AppilcationConstanst and MessageResponse

+++ src/main/java/com/hcmute/hotel/controller/ProvinceController.java
@@ -1,13 +1,17 @@
 package com.hcmute.hotel.controller;
 
+import com.hcmute.hotel.constants.ApplicationConstants;
 import com.hcmute.hotel.mapping.ProvinceMapping;
 import com.hcmute.hotel.model.entity.ProvinceEntity;
 import com.hcmute.hotel.model.entity.StayEntity;
-import com.hcmute.hotel.model.payload.SuccessResponse;
 import com.hcmute.hotel.model.payload.request.Province.AddNewProvinceRequest;
 import com.hcmute.hotel.model.payload.request.Province.UpdateProvinceRequest;
+import com.hcmute.hotel.model.payload.response.DataResponse;
+import com.hcmute.hotel.model.payload.response.MessageResponse;
 import com.hcmute.hotel.security.JWT.JwtUtils;
 import com.hcmute.hotel.service.ProvinceService;
+import io.swagger.annotations.ApiOperation;
+import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.ComponentScan;
@@ -20,7 +24,10 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
+import javax.websocket.OnClose;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static com.google.common.net.HttpHeaders.AUTHORIZATION;
 
@@ -34,49 +41,33 @@ public class ProvinceController {
     @Autowired
     JwtUtils jwtUtils;
 
-    @GetMapping("/getall")
-    public ResponseEntity<SuccessResponse> getAllProvinces() {
+    @GetMapping("")
+    @ApiOperation("Find all")
+    public DataResponse getAllProvinces() {
         List<ProvinceEntity> provinceEntityList = provinceService.getAllProvinces();
-        if (provinceEntityList.size() == 0) {
-            SuccessResponse response = new SuccessResponse();
-            response.setStatus(HttpStatus.FOUND.value());
-            response.setMessage("List province is empty");
-            response.setSuccess(false);
-            return new ResponseEntity<>(response, HttpStatus.FOUND);
-        }
-        SuccessResponse response = new SuccessResponse();
-        response.setStatus(HttpStatus.OK.value());
-        response.setMessage("Query Successfully");
-        response.setSuccess(true);
-        response.getData().put("List province:", provinceEntityList);
-        return new ResponseEntity<>(response, HttpStatus.OK);
+        return new DataResponse(provinceEntityList);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<SuccessResponse> getProvinceById(@PathVariable int id) {
+    @ApiOperation("Find by id")
+    public ResponseEntity<Object> getProvinceById(@PathVariable int id) {
         ProvinceEntity provinceEntity = provinceService.getProvinceById(id);
         if (provinceEntity == null) {
-            SuccessResponse response = new SuccessResponse();
-            response.setStatus(HttpStatus.FOUND.value());
-            response.setMessage("province is not found");
-            response.setSuccess(false);
-            return new ResponseEntity<>(response, HttpStatus.FOUND);
+            MessageResponse messageResponse = new MessageResponse("Bad Request", "PROVINCE_ID_NOT_FOUND", "Province id not found");
+            return new ResponseEntity<>(messageResponse, HttpStatus.NOT_FOUND);
         }
-        SuccessResponse response = new SuccessResponse();
-        response.setStatus(HttpStatus.OK.value());
-        response.setMessage("Query Successfully");
-        response.setSuccess(true);
-        response.getData().put("province", provinceEntity);
-        return new ResponseEntity<>(response, HttpStatus.OK);
+        Map<String, Object> map = new HashMap<>();
+        map.put("content", provinceEntity);
+        return new ResponseEntity<>(map, HttpStatus.OK);
     }
 
-    @PostMapping("/add")
-    public ResponseEntity<SuccessResponse> addProvince(@RequestBody @Valid AddNewProvinceRequest addNewProvinceRequest, BindingResult result, HttpServletRequest httpServletRequest) throws Exception {
+    @PostMapping("")
+    @ApiOperation("Create")
+    public ResponseEntity<Object> addProvince(@RequestBody @Valid AddNewProvinceRequest addNewProvinceRequest, BindingResult result, HttpServletRequest httpServletRequest) throws Exception {
         if (result.hasErrors()) {
-            throw new MethodArgumentNotValidException(null,result);
+            throw new MethodArgumentNotValidException(null, result);
         }
         String authorizationHeader = httpServletRequest.getHeader(AUTHORIZATION);
-        SuccessResponse response = new SuccessResponse();
         if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
             String accessToken = authorizationHeader.substring("Bearer ".length());
 
@@ -85,30 +76,28 @@ public class ProvinceController {
             }
             ProvinceEntity province = ProvinceMapping.addProvinceToEntity(addNewProvinceRequest);
             boolean isExisted = provinceService.findByName(addNewProvinceRequest.getName());
-            if(isExisted == false) {
-                provinceService.saveProvince(province);
-                response.setStatus(HttpStatus.OK.value());
-                response.setMessage("Save province successfully");
-                response.setSuccess(true);
-                response.getData().put("province", province);
-                return new ResponseEntity<>(response, HttpStatus.OK);
+            if (isExisted == false) {
+                province = provinceService.saveProvince(province);
+                Map<String, Object> map = new HashMap<>();
+                map.put("content", province);
+                return new ResponseEntity<>(map, HttpStatus.OK);
             }
             else {
-                response.setStatus(HttpStatus.BAD_REQUEST.value());
-                response.setMessage("province name is existed");
-                response.setSuccess(false);
-                return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-            }
-        } else throw new BadCredentialsException("access token is missing");
-    }
+                MessageResponse messageResponse = new MessageResponse("Bad Request", "PROVINCE_NAME_EXISTED", "Province name has been used");
+                return new ResponseEntity<>(messageResponse, HttpStatus.BAD_REQUEST);
+                }
+            } else throw new BadCredentialsException("access token is missing");
+        }
 
-    @PatchMapping("/update/{id}")
-    public ResponseEntity<SuccessResponse> updateProvince(@Valid @RequestBody UpdateProvinceRequest updateProvinceRequest, BindingResult result, HttpServletRequest httpServletRequest, @PathVariable("id") int id) throws Exception {
+
+
+    @PatchMapping("/{id}")
+    @ApiOperation("Update")
+    public ResponseEntity<Object> updateProvince(@Valid @RequestBody UpdateProvinceRequest updateProvinceRequest, BindingResult result, HttpServletRequest httpServletRequest, @PathVariable("id") int id) throws Exception {
         if (result.hasErrors()) {
-            throw new MethodArgumentNotValidException(null,result);
+            throw new MethodArgumentNotValidException(null, result);
         }
         String authorizationHeader = httpServletRequest.getHeader(AUTHORIZATION);
-        SuccessResponse response = new SuccessResponse();
         if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
             String accessToken = authorizationHeader.substring("Bearer ".length());
 
@@ -117,49 +106,45 @@ public class ProvinceController {
             }
             ProvinceEntity province = provinceService.getProvinceById(id);
             if (province == null) {
-                response.setStatus(HttpStatus.BAD_REQUEST.value());
-                response.setMessage("Can't find province with id " + id);
-                response.setSuccess(false);
-                return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+                MessageResponse messageResponse = new MessageResponse("Bad Request", "PROVINCE_ID_NOT_FOUND", "Province id not found");
+                return new ResponseEntity<>(messageResponse, HttpStatus.NOT_FOUND);
             }
             boolean isExisted = provinceService.findByNameAndId(updateProvinceRequest.getName(), id);
-            if(!isExisted) {
+            if (!isExisted) {
                 province = ProvinceMapping.updateProvinceToEntity(updateProvinceRequest, province);
-                provinceService.saveProvince(province);
-                response.setStatus(HttpStatus.OK.value());
-                response.setMessage("Update province successfully");
-                response.setSuccess(true);
-                response.getData().put("province", province);
-                return new ResponseEntity<>(response, HttpStatus.OK);
-            }
-            else {
-                response.setStatus(HttpStatus.BAD_REQUEST.value());
-                response.setMessage("province name is existed");
-                response.setSuccess(false);
-                return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+                province = provinceService.saveProvince(province);
+                Map<String, Object> map = new HashMap<>();
+                map.put("content", province);
+                return new ResponseEntity<>(map, HttpStatus.OK);
+            } else {
+                MessageResponse messageResponse = new MessageResponse("Bad Request", "PROVINCE_NAME_EXISTED", "Province name has been used");
+                return new ResponseEntity<>(messageResponse, HttpStatus.BAD_REQUEST);
             }
         } else throw new BadCredentialsException("access token is missing");
     }
 
-    @DeleteMapping("/delete/{id}")
-    public ResponseEntity<SuccessResponse> deleteProvinceById(@PathVariable("id") int id, HttpServletRequest httpServletRequest) {
+    @DeleteMapping("/{id}")
+    @ApiOperation("Delete")
+    public ResponseEntity<Object> deleteProvinceById(@PathVariable("id") int id, HttpServletRequest httpServletRequest) {
         String authorizationHeader = httpServletRequest.getHeader(AUTHORIZATION);
-        ProvinceEntity province=provinceService.getProvinceById(id);
-        SuccessResponse response = new SuccessResponse();
+        ProvinceEntity province = provinceService.getProvinceById(id);
         if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
             String accessToken = authorizationHeader.substring("Bearer ".length());
             if (jwtUtils.validateExpiredToken(accessToken) == true) {
                 throw new BadCredentialsException("access token is  expired");
             }
-            for (StayEntity stay : province.getStay())
-            {
-                stay.setProvince(null);
+            ProvinceEntity provinceEntity = provinceService.getProvinceById(id);
+            if (provinceEntity == null) {
+                MessageResponse messageResponse = new MessageResponse("Bad Request", "PROVINCE_ID_NOT_FOUND", "Province id not found");
+                return new ResponseEntity<>(messageResponse, HttpStatus.NOT_FOUND);
             }
-            provinceService.deleteById(id);
-            response.setStatus(HttpStatus.OK.value());
-            response.setMessage("Delete province successfully");
-            response.setSuccess(true);
-            return new ResponseEntity<>(response,HttpStatus.OK);
+            else {
+                for (StayEntity stay : province.getStay()) {
+                    stay.setProvince(null);
+                }
+                provinceService.deleteById(id);
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
         } else throw new BadCredentialsException("access token is missing");
     }
 }
